requirements.txt
openai
streamlit
chromadb
sentence-transformers
python-dotenv
nltk
tqdm
